# backend/src/main/resources/openapi/swagger.yaml
swagger: "2.0"
info:
  description: "Gestion des automobiles de collection"
  version: "1.0.0"
  title: "Gestion des automobiles de collection"
  contact:
    email: "apiteam@swagger.io"
basePath: "/api"
produces:
  - application/json
tags:
  - name: "car"
    description: "gestion des voiture"
  - name: "repair"
    description: "Réparations de la voiture"
  - name: "user"
    description: "partie connexion - creation compte"
schemes:
  - "https"
  - "http"
paths:
  /car:
    post:
      tags:
        - "car"
      summary: "Ajout d'une nouvelle voiture"
      description: ""
      operationId: "addCar"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Car object that needs to be added to the store"
          required: true
          schema:
            $ref: "#/definitions/Car"
      responses:
        "201":
          description: "voiture ajoutée"
        "405":
          description: "Invalid input"
  /car/{carId}:
    get:
      tags:
        - "car"
      summary: "Trouver une voiture par ID"
      description: "Returns a single car"
      operationId: "getCarById"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Car"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Car not found"
    patch:
      tags:
        - "car"
      summary: "Modifier une voiture existante"
      description: ""
      operationId: "updateCar"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: params
          required: true
          schema:
            $ref: "#/definitions/Car"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Car"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Car not found"
        "405":
          description: "Validation exception"
    delete:
      tags:
        - "car"
      summary: "Supprimer une voiture"
      description: ""
      operationId: "deleteCar"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: string
            example: "OK"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Car not found"
    parameters:
      - name: "carId"
        in: "path"
        description: "Car id"
        required: true
        type: "string"
        format: "uuid"
  /car/gellAll/{userId}:
    get:
      tags:
        - "car"
      summary: "Trouver toute les voitures"
      description: "Returns all cars"
      operationId: "getCarsByUserId"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/Car"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Car not found"
    parameters:
      - name: "userId"
        in: "path"
        description: "User id"
        required: true
        type: "string"
        format: "uuid"
  /user/authentification:
    post:
      tags:
        - "user"
      summary: "connexion"
      description: "connexion via user + passwd"
      operationId: "connexion"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/Authentification"
      responses:
        "200":
          description: "user connected"
          schema:
            $ref: "#/definitions/UserInfos"
        "404":
          description: "User not found"
  /user/create:
    post:
      tags:
        - "user"
      summary: "creation"
      description: "creation de compte"
      operationId: "create"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/UserInfos"
      responses:
        "201":
          description: "user created"
        "404":
          description: "User not found"
  /user/modification:
    patch:
      tags:
        - "user"
      summary: "modification"
      description: "modification de compte"
      operationId: "modification"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/UserInfos"
      responses:
        "200":
          description: "user modified"
          schema:
            type: string
            example: "OK"
        "404":
          description: "User not found"
  /repair:
    post:
      tags:
        - "repair"
      summary: "reparation"
      description: "ajout d'une reparation"
      operationId: "addRepair"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/Repair"
      responses:
        "201":
          description: "Réparation ajoutée"
        "404":
          description: "Réparation KO"
  /repair/{reparationId}:
    delete:
      tags:
        - "repair"
      summary: "Supprimer une reparation"
      description: ""
      operationId: "deleteRepair"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: string
            example: "OK"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Repair not found"
    parameters:
      - name: "reparationId"
        in: "path"
        description: "Repair id"
        required: true
        type: "string"
        format: "uuid"
  /repair/gellAll/{voitureId}:
    get:
      tags:
        - "repair"
      summary: "Trouver toute les réparations d'une voiture"
      description: "Returns all repairs"
      operationId: "getRepairByCarId"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/Repair"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Car not found"
    parameters:
      - name: "voitureId"
        in: "path"
        description: "Car id"
        required: true
        type: "string"
        format: "uuid"
definitions:
  Car:
    type: "object"
    properties:
      marque:
        type: "string"
        example: "Peugeot"
      modele:
        type: "string"
        example: "205 GTI"
      premiereImmat:
        type: "string"
        format: "date"
        example: "10-10-1991"
      puissanceFiscale:
        type: "integer"
        example: "8"
      prixDachat:
        type: "integer"
        example: "10000"
      prixVenteEstimee:
        type: "integer"
        example: "15000"
      userId:
        type: "string"
        format: "uuid"
        example: "0000-0000"
      id:
        type: "string"
        format: "uuid"
        example: "0000-0000"
  Authentification:
    type: "object"
    properties:
      login:
        type: "string"
        example: "toto"
      password:
        type: "string"
        example: "tata"
  UserInfos:
    type: "object"
    properties:
      login:
        type: "string"
        example: "toto"
      password:
        type: "string"
        example: "tata"
      prenom:
        type: "string"
        example: "Jacques"
      nom:
        type: "string"
        example: "Martin"
      id:
        type: "string"
        format: "uuid"
        example: "0000-0000"
  Repair:
    type: "object"
    properties:
      id:
        type: "string"
        format: "uuid"
        example: "0000-0000"
      voiture_id:
        type: "string"
        format: "uuid"
        example: "0000-0000"
      type:
        type: "string"
        example: "freins"
      facture:
        type: "boolean"
        example: "false"